feat: writing to a file
