+++ src/app.ts
@@ -4,7 +4,7 @@ import fastify from 'fastify'
 import boardRouter from './resources/board/board.router'
 import tasksRouter from './resources/tasks/task.router'
 import usersRouter from './resources/users/user.router'
-import { Logger } from './logger'
+import { Logger } from './log/logger'
 
 const app = fastify()
